+++ dlls/ntdll/unix/tape.c
@@ -66,7 +66,7 @@
 
 WINE_DEFAULT_DEBUG_CHANNEL(tape);
 
-static const char *io2str( DWORD io )
+static const char *io2str( unsigned int io )
 {
     switch (io)
     {
@@ -109,7 +109,7 @@ static NTSTATUS TAPE_CreatePartition( int fd, const TAPE_CREATE_PARTITION *data 
     struct mtop cmd;
 
     TRACE( "fd: %d method: 0x%08x count: 0x%08x size: 0x%08x\n",
-           fd, data->Method, data->Count, data->Size );
+           fd, (int)data->Method, (int)data->Count, (int)data->Size );
 
     if (data->Count > 1)
     {
@@ -131,7 +131,7 @@ static NTSTATUS TAPE_CreatePartition( int fd, const TAPE_CREATE_PARTITION *data 
         break;
 #endif
     default:
-        ERR( "Unhandled method: 0x%08x\n", data->Method );
+        ERR( "Unhandled method: 0x%08x\n", (int)data->Method );
         return STATUS_INVALID_PARAMETER;
     }
 
@@ -151,7 +151,7 @@ static NTSTATUS TAPE_Erase( int fd, const TAPE_ERASE *data )
     struct mtop cmd;
 
     TRACE( "fd: %d type: 0x%08x immediate: 0x%02x\n",
-           fd, data->Type, data->Immediate );
+           fd, (int)data->Type, (int)data->Immediate );
 
     switch (data->Type)
     {
@@ -164,7 +164,7 @@ static NTSTATUS TAPE_Erase( int fd, const TAPE_ERASE *data )
         cmd.mt_count = 0;
         break;
     default:
-        ERR( "Unhandled type: 0x%08x\n", data->Type );
+        ERR( "Unhandled type: 0x%08x\n", (int)data->Type );
         return STATUS_INVALID_PARAMETER;
     }
 
@@ -253,7 +253,7 @@ static NTSTATUS TAPE_GetMediaParams( int fd, TAPE_GET_MEDIA_PARAMETERS *data )
 /******************************************************************
  *      TAPE_GetPosition
  */
-static NTSTATUS TAPE_GetPosition( int fd, ULONG type, TAPE_GET_POSITION *data )
+static NTSTATUS TAPE_GetPosition( int fd, UINT type, TAPE_GET_POSITION *data )
 {
 #ifdef HAVE_SYS_MTIO_H
     struct mtget get;
@@ -312,7 +312,7 @@ static NTSTATUS TAPE_Prepare( int fd, const TAPE_PREPARE *data )
     struct mtop cmd;
 
     TRACE( "fd: %d type: 0x%08x immediate: 0x%02x\n",
-           fd, data->Operation, data->Immediate );
+           fd, (int)data->Operation, (int)data->Immediate );
 
     switch (data->Operation)
     {
@@ -345,7 +345,7 @@ static NTSTATUS TAPE_Prepare( int fd, const TAPE_PREPARE *data )
         /* Native ignores this if the drive doesn't support it */
         return STATUS_SUCCESS;
     default:
-        ERR( "Unhandled operation: 0x%08x\n", data->Operation );
+        ERR( "Unhandled operation: 0x%08x\n", (int)data->Operation );
         return STATUS_INVALID_PARAMETER;
     }
 
@@ -367,7 +367,7 @@ static NTSTATUS TAPE_SetDriveParams( int fd, const TAPE_SET_DRIVE_PARAMETERS *da
     TRACE( "fd: %d ECC: 0x%02x, compression: 0x%02x padding: 0x%02x\n",
             fd, data->ECC, data->Compression, data->DataPadding );
     TRACE( "setmarks: 0x%02x zonesize: 0x%08x\n",
-           data->ReportSetmarks, data->EOTWarningZoneSize );
+           (int)data->ReportSetmarks, (int)data->EOTWarningZoneSize );
 
     if (data->ECC || data->DataPadding || data->ReportSetmarks ||
         data->EOTWarningZoneSize ) WARN( "Setting not supported\n" );
@@ -390,7 +390,7 @@ static NTSTATUS TAPE_SetMediaParams( int fd, const TAPE_SET_MEDIA_PARAMETERS *da
 #ifdef HAVE_SYS_MTIO_H
     struct mtop cmd;
 
-    TRACE( "fd: %d blocksize: 0x%08x\n", fd, data->BlockSize );
+    TRACE( "fd: %d blocksize: 0x%08x\n", fd, (int)data->BlockSize );
 
     cmd.mt_op = MTSETBLK;
     cmd.mt_count = data->BlockSize;
@@ -411,7 +411,7 @@ static NTSTATUS TAPE_SetPosition( int fd, const TAPE_SET_POSITION *data )
     struct mtop cmd;
 
     TRACE( "fd: %d method: 0x%08x partition: 0x%08x offset: 0x%s immediate: 0x%02x\n",
-           fd, data->Method, data->Partition, wine_dbgstr_longlong(data->Offset.QuadPart),
+           fd, (int)data->Method, (int)data->Partition, wine_dbgstr_longlong(data->Offset.QuadPart),
            data->Immediate );
 
     if (sizeof(cmd.mt_count) < sizeof(data->Offset.QuadPart) &&
@@ -467,7 +467,7 @@ static NTSTATUS TAPE_SetPosition( int fd, const TAPE_SET_POSITION *data )
         WARN( "Positioning method not supported\n" );
         return STATUS_INVALID_PARAMETER;
     default:
-        ERR( "Unhandled method: 0x%08x\n", data->Method );
+        ERR( "Unhandled method: 0x%08x\n", (int)data->Method );
         return STATUS_INVALID_PARAMETER;
     }
 
@@ -487,7 +487,7 @@ static NTSTATUS TAPE_WriteMarks( int fd, const TAPE_WRITE_MARKS *data )
     struct mtop cmd;
 
     TRACE( "fd: %d type: 0x%08x count: 0x%08x immediate: 0x%02x\n",
-           fd, data->Type, data->Count, data->Immediate );
+           fd, (int)data->Type, (int)data->Count, data->Immediate );
 
     switch (data->Type)
     {
@@ -504,7 +504,7 @@ static NTSTATUS TAPE_WriteMarks( int fd, const TAPE_WRITE_MARKS *data )
         cmd.mt_count = data->Count;
         break;
     default:
-        ERR( "Unhandled type: 0x%08x\n", data->Type );
+        ERR( "Unhandled type: 0x%08x\n", (int)data->Type );
         return STATUS_INVALID_PARAMETER;
     }
 
@@ -519,8 +519,8 @@ static NTSTATUS TAPE_WriteMarks( int fd, const TAPE_WRITE_MARKS *data )
  *		tape_DeviceIoControl
  */
 NTSTATUS tape_DeviceIoControl( HANDLE device, HANDLE event, PIO_APC_ROUTINE apc, void *apc_user,
-                               client_ptr_t io, ULONG code,
-                               void *in_buffer, ULONG in_size, void *out_buffer, ULONG out_size )
+                               client_ptr_t io, UINT code,
+                               void *in_buffer, UINT in_size, void *out_buffer, UINT out_size )
 {
     DWORD sz = 0;
     NTSTATUS status = STATUS_INVALID_PARAMETER;
